refactor(entities/comment): 타입 구조를 request/response로 분리
- CommentFormData, CommentRequest를 request.types.ts로 분리
- Comment, CommentsResponse를 response.types.ts로 분리
- 기존 types.ts 파일 제거
- Post 도메인과 동일한 타입 구조로 통일
